chore: rename virtual modules
