--- packages/vite-glob-routes/src/virtual.d.ts
@@ -1,4 +1,0 @@
-declare module "virtual:api-routes/hattip" {
-  const value: import("@hattip/compose").RequestHandler;
-  export default value;
-}
