+++ packages/vite-glob-routes/types/virtual.d.ts
@@ -1,9 +1,9 @@
-declare module "virtual:page-routes/react-router" {
+declare module "virtual:glob-page-routes/react-router" {
   const value: import("react-router").RouteObject[];
   export default value;
 }
 
-declare module "virtual:api-routes/hattip" {
+declare module "virtual:glob-api-routes/hattip" {
   const value: import("@hattip/compose").RequestHandler;
   export default value;
 }
