+++ src/client/index.tsx
@@ -1,5 +1,5 @@
 import "virtual:uno.css";
-import globPageRoutes from "virtual:page-routes/react-router";
+import globPageRoutes from "virtual:glob-page-routes/react-router";
 import { tinyassert } from "@hiogawa/utils";
 import React from "react";
 import { createRoot } from "react-dom/client";
